ref: Minimize amount of code in main function
